pythonize
